Move ReplaceOrAppendEnvValues to container package
